--- JSON/subscriptions.json
@@ -1,2 +1,0 @@
-{
-}
